Set default project on project dependency spec
By moving `DependencySpecContainer` creation knowlege to a factory
service so the dependency on `ProjectIdentifier` can be properly
declared at that level.

The functional benefit is that `ProjectDependencySpec` resolution is no
longer context dependent and instances can be safely passed around as
their meaning is now unambiguous, they always represent a dependency on
a particular component of a specific project.

The structural benefit is the confinement of the `project path` /
`ProjectIdentifier` dependency to the single place that actually
requires it, the instantiation logic of
`DefaultDependencySpecContainer`. This has a cascade effect on the call
hierarchy of `newDefaultLibraryLocalComponentMetadata` which no longer
has to thread the project path argument through.

This commit assumes constructor injection for default model element
implementations is supported. Wouldn't it be nice?